fix(protocol): unify chain_id to u32 (Mainnet=1001) and align canonical tx schema/docs (ISSUE-002, ISSUE-003)
- Seragamkan chain_id u64->u32 pada runtime, transport, handshake (188B->184B), snapshot, telemetry, health, explorer, rpc, dan DTO dispatcher.

- Hapus default divergen (transport/wallet 1 -> GENESIS_CHAIN_ID); hapus cast as u64.

- Dokumen konstitusi/operasional disinkronkan ke Header 8B, preimage aktual, payload 24 KiB; reference vector diregenerasi.

- Ganti hardcode 148 di mempool dengan TRANSACTION_BASE_BYTES + 4 + payload.

- Tambah tests/golden_vectors.rs (byte-level golden lintas wallet/validator).

- Verifikasi: cargo test --workspace hijau, guardrail 100%.

+++ src/platform/cli/dispatcher.rs
@@ -237,7 +237,7 @@ struct SnapshotMetadataInfo {
     file_path: String,
     magic: String,
     version: u32,
-    chain_id: u64,
+    chain_id: u32,
     height: u64,
     epoch: u64,
     block_hash: String,
@@ -269,7 +269,7 @@ struct FaucetRequestInfo {
 #[derive(Serialize)]
 struct ExplorerSummaryInfo {
     network: &'static str,
-    chain_id: u64,
+    chain_id: u32,
     current_height: u64,
     finalized_height: u64,
     mempool_size: usize,
@@ -280,7 +280,7 @@ struct ExplorerSummaryInfo {
 #[derive(Serialize)]
 struct NetworkStatusInfo {
     network: &'static str,
-    chain_id: u64,
+    chain_id: u32,
     genesis_block_hash: String,
     state_root: String,
     p2p_wire_magic: &'static str,
@@ -302,7 +302,7 @@ struct NetworkPeersInfo {
 struct NodeLaunchStatusInfo {
     network: &'static str,
     role: &'static str,
-    chain_id: u64,
+    chain_id: u32,
     genesis_block_hash: String,
     state_root: String,
     storage_engine: &'static str,
@@ -2357,7 +2357,7 @@ pub async fn dispatch(command: CliCommand, format: OutputFormat) -> Result<(), S
                     let genesis = CeremonyTranscript::canonical_mainnet_genesis();
                     let info = NetworkStatusInfo {
                         network: "aurion-mainnet",
-                        chain_id: crate::genesis::builder::GENESIS_CHAIN_ID as u64,
+                        chain_id: crate::genesis::builder::GENESIS_CHAIN_ID,
                         genesis_block_hash: genesis.header.compute_block_hash().to_hex(),
                         state_root: genesis.header.state_root.to_hex(),
                         p2p_wire_magic: "AUR0",
